+++ src/interfaces/IConfigType.ts
@@ -1,0 +1,6 @@
+/** configuration source type */
+export enum IConfigType {
+	File = 1,
+	DotEnv = 2,
+	Env = 3,
+}